+++ shufflebot/src/commands/banUser.js
