+++ shufflebot/src/commands/kickUser.js
