+++ shufflebot/src/commands/removeBlooWord.js
@@ -1,16 +1,14 @@
 const {ApplicationCommandOptionType, PermissionFlagsBits, EmbedBuilder, AttachmentBuilder, Client, Interaction, Attachment, SlashCommandBuilder} = require("discord.js");
-const bloosinferno = require('../../models/bloosinferno')
-
+const bloosinferno = require('../models/bloosinferno')
 module.exports = {
     data: new SlashCommandBuilder()
-        .setName('addblooword')
-        .setDescription('add bloo word')
+        .setName('removeblooword')
+        .setDescription('heretic heretic heretic heretic heretic heretic heretic heretic heretic heretic heretic heretic')
         .addStringOption(option => 
             option.setName('word')
-                .setDescription('bloo word')
+                .setDescription('word of bloo')
                 .setRequired(true)
         ),
-
 
         /**
      * 
@@ -29,13 +27,13 @@ module.exports = {
             const oldWord = oldMap.words.includes(interaction.options.get('word').value)
             const hasRole = interaction.member.roles.cache.some(r => r.name === 'bailiff')
             const options = { upsert : false }
-            if (!oldWord && hasRole) {
+            if (oldWord && hasRole) {
                 interaction.editReply('⠀')
-                var votes = oldMap.words.push(word)
+                var votes = oldMap.words.splice(oldMap.words.indexOf(oldWord), 1)
                 oldMap.save(votes)
             } else {
                 interaction.editReply(
-                    'That word does already exist or you cant do that'
+                    'That word doesnt exist or you cant do that'
                 )
             }
         } catch (error) {
